+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   [Ken Grigsby](https://github.com/kgrigsby59), [#112](https://github.com/AliSoftware/SwiftGen/issues/112)
 
 
+## Master
+
+* Added support for TXT (`colors.txt`) files to have named value.  
+[Derek Ostrander](https://github.com/dostrander)
+
 ## 0.8.0
 
 * Introducing alternative way to install SwiftGen: using CocoaPods! See README for more details.